Added Delta time to the Blue Machine

+++ source/Class/boss.py
@@ -397,7 +397,8 @@ class Boss:
                 self.boss.shape("Textures/Enemies/Boss_Scaled.gif")
             else:
                 self.boss.shape("Textures/Enemies/Boss.gif")
-            self.boss.goto(random.randint(-640 * scale_factor_x, 640 * scale_factor_x), random.randint(120 * scale_factor_y, 220 * scale_factor_y))
+            # Want to cast these ranges to integers to avoid a crash at certain resolutions
+            self.boss.goto(random.randint(int(-640 * scale_factor_x), int(640 * scale_factor_x)), random.randint(int(120 * scale_factor_y), int(220 * scale_factor_y)))
             self.update = 3.5
             return
