+++ source/Class/redMachine.py
@@ -445,7 +445,8 @@ class RedMachine:
                 self.red_machine.shape("Textures/Enemies/Enemy(11-15)_Scaled.gif")
             else:
                 self.red_machine.shape("Textures/Enemies/Enemy(11-15).gif")
-            self.red_machine.goto(random.randint(-640 * scale_factor_x, 640 * scale_factor_x), random.randint(120 * scale_factor_y, 220 * scale_factor_y))
+            # Want to cast these ranges to integers to avoid a crash at certain resolutions
+            self.red_machine.goto(random.randint(int(-640 * scale_factor_x), int(640 * scale_factor_x)), random.randint(int(120 * scale_factor_y), int(220 * scale_factor_y)))
             self.update = 3.5
             return
